SWDEV-272811 - hipCabs returns a double and uses sqrtf, which maps to f32 operation. sqrt should be used instead


[ROCm/hip commit: 4d0383df0c8806a0ff84b164c500466094398353]

+++ include/hip/amd_detail/hip_complex.h
@@ -214,7 +214,7 @@ __device__ __host__ static inline hipDoubleComplex hipCdiv(hipDoubleComplex p, h
     return ret;
 }
 
-__device__ __host__ static inline double hipCabs(hipDoubleComplex z) { return sqrtf(hipCsqabs(z)); }
+__device__ __host__ static inline double hipCabs(hipDoubleComplex z) { return sqrt(hipCsqabs(z)); }
 
 
 #if __cplusplus
